lab-6/requirements.txt
googleapis-common-protos
google-cloud-storage
google-cloud-pubsub
